requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
audioop-lts==0.2.2
certifi==2025.10.5
charset-normalizer==3.4.4
coc.py==3.9.1
discord==2.3.2
discord.py==2.6.4
dotenv==0.9.9
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
orjson==3.11.5
propcache==0.4.1
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.5.0
yarl==1.22.0
